requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
datadog==0.47.0
idna==3.6
influxdb-client==1.41.0
python-dateutil==2.9.0.post0
pytz==2024.1
reactivex==4.0.4
requests==2.31.0
six==1.16.0
typing_extensions==4.10.0
urllib3==2.1.0
